fix(web): Keep usage meter UI off Convex server modules
Vite evaluates the full rateLimits import graph in dev, so settings hit process.env from generated server.js and crashed with a 500.
