+++ apps/web/src/convex/lib/rateLimits.ts
@@ -20,34 +20,27 @@ import {
 	type SupportedServiceTier
 } from '@convex/lib/models';
 import { ensureSubscription, tierLimits, type TierLimits } from '@convex/lib/tiers';
+import {
+	URL_SCRAPE_USAGE_UNITS,
+	WEB_SEARCH_USAGE_UNITS,
+	usageMeters,
+	usagePeriods,
+	type UsageMeterId,
+	type UsagePeriod
+} from '@convex/lib/usageMeters';
 import { vModelId, vServiceTier } from '@convex/lib/validators';
 
+export {
+	URL_SCRAPE_USAGE_UNITS,
+	WEB_SEARCH_USAGE_UNITS,
+	usageMeters,
+	usagePeriods,
+	type UsageMeterId,
+	type UsagePeriod
+};
+
 const MONTH = 30 * DAY;
-/** Flat web-tools meter cost for one URL scrape. */
-export const URL_SCRAPE_USAGE_UNITS = 1.5;
-/** Flat web-tools meter cost for one web search (independent of result count). */
-export const WEB_SEARCH_USAGE_UNITS = 7;
 export const rateLimiter = new RateLimiter(components.rateLimiter, {});
-
-export const usageMeters = [
-	{
-		id: 'modelUsage',
-		label: 'Model usage',
-		noun: 'model usage',
-		description: 'More expensive models use up more usage.'
-	},
-	{
-		id: 'webTools',
-		label: 'Web tools',
-		noun: 'web tools',
-		description: 'Web search and URL scrape share this quota.'
-	}
-] as const;
-
-export type UsageMeterId = (typeof usageMeters)[number]['id'];
-
-export const usagePeriods = ['weekly', 'monthly'] as const;
-export type UsagePeriod = (typeof usagePeriods)[number];
 
 const periodDurations: Record<UsagePeriod, number> = { weekly: WEEK, monthly: MONTH };
